Migrate away from event loop to locking.
